Remove explicit auth mechanism and implement it through proxy config and secret
