+++ clickhouse/src/datasources/click-house-datasource/click-house-datasource-types.ts
@@ -17,8 +17,6 @@ import { DatasourceClient } from '@perses-dev/plugin-system';
 export interface ClickHouseDatasourceSpec {
   directUrl?: string;
   proxy?: HTTPProxy;
-  username?: string;
-  password?: string;
 }
 
 interface QueryRequestParameters extends Record<string, string> {
