+++ clickhouse/src/model/click-house-client.ts
@@ -11,6 +11,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { RequestHeaders } from '@perses-dev/core';
+
+export interface ClickHouseQueryParams {
+  query: string;
+  database?: string;
+}
+
+export interface ClickHouseQueryOptions {
+  datasourceUrl: string;
+  headers?: RequestHeaders;
+}
+
 export interface ClickHouseQueryResponse {
   status: 'success' | 'error';
   data: any;
@@ -18,4 +30,61 @@ export interface ClickHouseQueryResponse {
 
 export interface ClickHouseClient {
   query: (params: { start: string; end: string; query: string }) => Promise<ClickHouseQueryResponse>;
+}
+
+export async function query(
+  params: ClickHouseQueryParams,
+  queryOptions: ClickHouseQueryOptions
+): Promise<ClickHouseQueryResponse> {
+  const { datasourceUrl, headers } = queryOptions;
+
+  const url = urlBuilder(datasourceUrl);
+  if (!params.query) {
+    throw new Error('No query provided in params');
+  }
+
+  let finalQuery = params.query.trim();
+  if (!finalQuery.toUpperCase().includes('FORMAT')) {
+    finalQuery += ' FORMAT JSON';
+  }
+
+  url.searchParams.set('query', finalQuery);
+  url.searchParams.set('database', params.database || 'default');
+
+  const init = {
+    method: 'GET',
+    headers: {
+      ...headers,
+    },
+  };
+
+  try {
+    const response = await fetch(url.toString(), init);
+
+    if (!response.ok) {
+      const errorText = await response.text();
+      console.error('ClickHouse error response:', errorText);
+      return {
+        status: 'error',
+        data: [],
+      };
+    }
+
+    const body = await response.json();
+
+    return {
+      status: 'success',
+      data: body.data || body,
+    };
+  } catch (e) {
+    throw new Error(`ClickHouse query failed: ${e}`);
+  }
+}
+
+function urlBuilder(datasourceUrl: string): URL {
+  if (datasourceUrl.startsWith('http://') || datasourceUrl.startsWith('https://')) {
+    return new URL(datasourceUrl);
+  }
+  // if relative path (e.g. proxy url), resolve against window.location.origin
+  return new URL(datasourceUrl, window.location.origin);
 }
